hospital_appointment_system/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
django-redis>=5.4.0
redis>=5.0.1
django-fernet-fields>=0.6
python-dotenv>=1.0.0
gunicorn>=21.2.0
